+++ elastic-job-cloud/elastic-job-cloud-scheduler/src/main/java/com/dangdang/ddframe/job/cloud/scheduler/context/JobContext.java
@@ -18,6 +18,7 @@
 package com.dangdang.ddframe.job.cloud.scheduler.context;
 
 import com.dangdang.ddframe.job.cloud.scheduler.config.CloudJobConfiguration;
+import com.dangdang.ddframe.job.context.ExecutionType;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;