+++ elastic-job-common/elastic-job-common-core/src/main/java/com/dangdang/ddframe/job/context/ExecutionType.java
@@ -15,14 +15,22 @@
  * </p>
  */
 
-package com.dangdang.ddframe.job.cloud.scheduler.context;
+package com.dangdang.ddframe.job.context;
 
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-        JobContextTest.class
-    })
-public final class AllContextTests {
+/**
+ * 执行类型.
+ *
+ * @author zhangliang
+ */
+public enum ExecutionType {
+    
+    /**
+     * 准备执行的任务.
+     */
+    READY,
+    
+    /**
+     * 失效转移的任务.
+     */
+    FAILOVER
 }